feature: add permission

+++ src/api.py
@@ -1,14 +1,18 @@
 import json
+from os import abort
 from flask import Flask, request
 from flask_cors import CORS
 from bson import ObjectId
+import flask_login
+import permission
 import requests
-from Util.db import rule_db, keywords_db, bili_mtr_db
+from Util.db import rule_db, keywords_db, bili_mtr_db, user_db
 from rule import keywords, update_keywords_list, update_rules
 from qqbot import send, get
-from user import register_user_module
+from user import current_login_user, register_user_module
 from flask_login import login_required
 import secrets
+from flask_login import current_user
 
 app = Flask(__name__)
 app.secret_key = secrets.token_urlsafe(16)
@@ -19,7 +23,9 @@ CORS(app)
 @app.route("/api/rules/", methods=["POST"], strict_slashes=False)
 @login_required
 def add_postprocess_file():
-    id = str(rule_db.insert_one(request.json).inserted_id)
+    r = request.json
+    r["creator"] = current_user.username
+    id = str(rule_db.insert_one(r).inserted_id)
     update_rules()
     return json.dumps({"_id": id})
 
@@ -27,7 +33,9 @@ def add_postprocess_file():
 @app.route("/api/rules/<rule_id>", methods=["POST"], strict_slashes=False)
 @login_required
 def update_postprocess_file(rule_id):
-    rule_db.update_one({"_id": ObjectId(rule_id)}, {"$set": request.json})
+    r = request.json
+    r["creator"] = current_user.username
+    rule_db.update_one({"_id": ObjectId(rule_id)}, {"$set": r})
     update_rules()
     return json.dumps({"status": "success"})
 
@@ -35,7 +43,7 @@ def update_postprocess_file(rule_id):
 @app.route("/api/rules/<rule_id>", methods=["DELETE"], strict_slashes=False)
 @login_required
 def delete_postprocess_file(rule_id):
-    rule_db.delete_one({"_id": ObjectId(rule_id)})
+    rule_db.delete_one({"_id": ObjectId(rule_id), "creator": current_user.username})
     update_rules()
     return json.dumps({"status": "success"})
 
@@ -44,7 +52,7 @@ def delete_postprocess_file(rule_id):
 @login_required
 def get_postprocess_file():
     res = []
-    for item in rule_db.find():
+    for item in rule_db.find({"creator": current_user.username}):
         item["_id"] = str(item["_id"])
         res.append(item)
 
@@ -55,7 +63,7 @@ def get_postprocess_file():
 @login_required
 def get_keywords():
     res = []
-    for item in keywords_db.find():
+    for item in keywords_db.find({"creator": current_user.username}):
         item["_id"] = str(item["_id"])
         res.append(item)
 
@@ -65,7 +73,9 @@ def get_keywords():
 @app.route("/api/keywords/", methods=["POST"], strict_slashes=False)
 @login_required
 def add_keywords():
-    id = str(keywords_db.insert_one(request.json).inserted_id)
+    r = request.json
+    r["creator"] = current_user.username
+    id = str(keywords_db.insert_one(r).inserted_id)
     update_keywords_list()
     return json.dumps({"_id": id})
 
@@ -73,7 +83,9 @@ def add_keywords():
 @app.route("/api/keywords/<keyword_id>", methods=["POST"], strict_slashes=False)
 @login_required
 def update_keywords(keyword_id):
-    keywords_db.update_one({"_id": ObjectId(keyword_id)}, {"$set": request.json})
+    r = request.json
+    r["creator"] = current_user.username
+    keywords_db.update_one({"_id": ObjectId(keyword_id)}, {"$set": r})
     update_keywords_list()
     return json.dumps({"status": "success"})
 
@@ -81,7 +93,9 @@ def update_keywords(keyword_id):
 @app.route("/api/keywords/<keyword_id>", methods=["DELETE"], strict_slashes=False)
 @login_required
 def delete_keywords(keyword_id):
-    keywords_db.delete_one({"_id": ObjectId(keyword_id)})
+    keywords_db.delete_one(
+        {"_id": ObjectId(keyword_id), "creator": current_user.username}
+    )
     update_keywords_list()
     return json.dumps({"status": "success"})
 
@@ -118,21 +132,29 @@ def remote_send_personal_msg():
 @login_required
 def group_list():
     group = get("groupList")
-    return json.dumps(group["data"])
+    group = group["data"]
+    perm = permission.get_current_permission()
+    perm_group = set(perm["group"])
+    group = [g for g in group if g["id"] in perm_group]
+    return json.dumps(group)
 
 
 @app.route("/api/friends", methods=["GET"])
 @login_required
 def friend_list():
-    group = get("friendList")
-    return json.dumps(group["data"])
+    person = get("friendList")
+    person = person["data"]
+    perm = permission.get_current_permission()
+    perm_person = set(perm["person"])
+    group = [g for g in person if g["id"] in perm_person]
+    return json.dumps(group)
 
 
 @app.route("/api/bili_monitor", methods=["GET"])
 @login_required
 def get_bili_mtr():
     res = []
-    for item in bili_mtr_db.find():
+    for item in bili_mtr_db.find({"creator": current_user.username}):
         item["_id"] = str(item["_id"])
         res.append(item)
     return json.dumps(res)
@@ -141,14 +163,16 @@ def get_bili_mtr():
 @app.route("/api/bili_monitor/<rule_id>", methods=["POST"], strict_slashes=False)
 @login_required
 def update_bili_mtr(rule_id):
-    bili_mtr_db.update_one({"_id": ObjectId(rule_id)}, {"$set": request.json})
+    r = request.json
+    r["creator"] = current_user.username
+    bili_mtr_db.update_one({"_id": ObjectId(rule_id)}, {"$set": r})
     return json.dumps({"status": "success"})
 
 
 @app.route("/api/bili_monitor/<rule_id>", methods=["DELETE"], strict_slashes=False)
 @login_required
 def delete_bili_mtr(rule_id):
-    bili_mtr_db.delete_one({"_id": ObjectId(rule_id)})
+    bili_mtr_db.delete_one({"_id": ObjectId(rule_id), "creator": current_user.username})
     return json.dumps({"status": "success"})
 
 
@@ -156,6 +180,7 @@ def delete_bili_mtr(rule_id):
 @login_required
 def add_bili_mtr():
     rule = request.json
+    rule["creator"] = current_user.username
     bid = rule["uid"]
     user_info = requests.get(
         "https://api.bilibili.com/x/space/acc/info?mid={}&jsonp=jsonp".format(bid)
@@ -163,3 +188,29 @@ def add_bili_mtr():
     rule["name"] = user_info
     id = str(bili_mtr_db.insert_one(rule).inserted_id)
     return json.dumps({"_id": id})
+
+
+@app.route("/api/permission/", methods=["GET"])
+@login_required
+def get_user_permission():
+    user = flask_login.current_user
+    print(user)
+    perm = permission.get_permission(user.username)
+    perm_list = []
+    if perm["role"] == 0:
+        for user in user_db.find():
+            perm_list.append(permission.get_permission(user["username"]))
+    else:
+        perm_list.append(perm)
+    return json.dumps(perm_list)
+
+
+@app.route("/api/permission/<username>", methods=["POST"])
+@login_required
+def update_user_permission(username):
+    perm = permission.get_current_permission()
+    if perm["role"] != 0:
+        return json.dumps({"status": "error", "error": "no permission"}), 403
+    r = request.json
+    permission.update_permission(username, {"person": r["person"], "group": r["group"]})
+    return json.dumps({"status": "success"})
